fix!: messages view
KK-1017
